+++ internal/bot/components/bios/core.go
@@ -1,12 +1,9 @@
 package bios
 
 import (
-	"errors"
 	"fmt"
 	"github.com/codemicro/dgo-toolkit/route"
 	"github.com/codemicro/lgballtDiscordBot/internal/db"
-	"github.com/codemicro/lgballtDiscordBot/internal/pluralkit"
-	"github.com/rs/zerolog/log"
 )
 
 func (*Bios) setBioField(bdt *db.UserBio, fieldName, newValue string, isSysmate bool, ctx *route.MessageContext) error {
@@ -43,19 +40,8 @@ func (*Bios) setBioField(bdt *db.UserBio, fieldName, newValue string, isSysmate 
 		return err
 	}
 
-	// When a message is proxied by PluralKit, this will apply the reactions to the correct message (and not the one
-	// that's just been/just about to be deleted)
-	// TODO: Turn this into a standalone function
-	targetMessageID := ctx.Message.ID
-	pkMsg, err := pluralkit.MessageById(targetMessageID)
-	if err != nil && !errors.Is(err, pluralkit.ErrorMessageNotFound) {
-		log.Warn().Err(err).Send()
-	} else if pkMsg != nil {
-		targetMessageID = pkMsg.Id
-	}
-
 	// react to message with a check mark to signify it worked
-	err = ctx.Session.MessageReactionAdd(ctx.Message.ChannelID, targetMessageID, "✅")
+	err = ctx.Session.MessageReactionAdd(ctx.Message.ChannelID, ctx.Message.ID, "✅")
 	return err
 }
 
@@ -86,20 +72,9 @@ func (b *Bios) clearBioField(bdt *db.UserBio, fieldName string, ctx *route.Messa
 		return err
 	}
 
-	// When a message is proxied by PluralKit, this will apply the reactions to the correct message (and not the one
-	// that's just been/just about to be deleted)
-	// TODO: Turn this into a standalone function
-	targetMessageID := ctx.Message.ID
-	pkMsg, err := pluralkit.MessageById(targetMessageID)
-	if err != nil && !errors.Is(err, pluralkit.ErrorMessageNotFound) {
-		log.Warn().Err(err).Send()
-	} else if pkMsg != nil {
-		targetMessageID = pkMsg.Id
-	}
-
 	// react to message with a check mark to signify it worked
 	for _, v := range []string{"🗑", "✅"} {
-		err = ctx.Session.MessageReactionAdd(ctx.Message.ChannelID, targetMessageID, v)
+		err = ctx.Session.MessageReactionAdd(ctx.Message.ChannelID, ctx.Message.ID, v)
 		if err != nil {
 			return err
 		}